# swagger_server/models/create_user.py
# coding: utf-8

from __future__ import absolute_import
from datetime import date, datetime  # noqa: F401

from typing import List, Dict  # noqa: F401

from swagger_server.models.base_model_ import Model
from swagger_server import util


class CreateUser(Model):
    """NOTE: This class is auto generated by the swagger code generator program.

    Do not edit the class manually.
    """

    def __init__(self, first_name: str=None, last_name: str=None, username: str=None, email: str=None, password: str=None):  # noqa: E501
        """CreateUser - a model defined in Swagger

        :param first_name: The first_name of this CreateUser.  # noqa: E501
        :type first_name: str
        :param last_name: The last_name of this CreateUser.  # noqa: E501
        :type last_name: str
        :param username: The username of this CreateUser.  # noqa: E501
        :type username: str
        :param email: The email of this CreateUser.  # noqa: E501
        :type email: str
        :param password: The password of this CreateUser.  # noqa: E501
        :type password: str
        """
        self.swagger_types = {
            'first_name': str,
            'last_name': str,
            'username': str,
            'email': str,
            'password': str
        }

        self.attribute_map = {
            'first_name': 'firstName',
            'last_name': 'lastName',
            'username': 'username',
            'email': 'email',
            'password': 'password'
        }

        self._first_name = first_name
        self._last_name = last_name
        self._username = username
        self._email = email
        self._password = password

    @classmethod
    def from_dict(cls, dikt) -> 'CreateUser':
        """Returns the dict as a model

        :param dikt: A dict.
        :type: dict
        :return: The CreateUser of this CreateUser.  # noqa: E501
        :rtype: CreateUser
        """
        return util.deserialize_model(dikt, cls)

    @property
    def first_name(self) -> str:
        """Gets the first_name of this CreateUser.


        :return: The first_name of this CreateUser.
        :rtype: str
        """
        return self._first_name

    @first_name.setter
    def first_name(self, first_name: str):
        """Sets the first_name of this CreateUser.


        :param first_name: The first_name of this CreateUser.
        :type first_name: str
        """

        self._first_name = first_name

    @property
    def last_name(self) -> str:
        """Gets the last_name of this CreateUser.


        :return: The last_name of this CreateUser.
        :rtype: str
        """
        return self._last_name

    @last_name.setter
    def last_name(self, last_name: str):
        """Sets the last_name of this CreateUser.


        :param last_name: The last_name of this CreateUser.
        :type last_name: str
        """

        self._last_name = last_name

    @property
    def username(self) -> str:
        """Gets the username of this CreateUser.


        :return: The username of this CreateUser.
        :rtype: str
        """
        return self._username

    @username.setter
    def username(self, username: str):
        """Sets the username of this CreateUser.


        :param username: The username of this CreateUser.
        :type username: str
        """

        self._username = username

    @property
    def email(self) -> str:
        """Gets the email of this CreateUser.


        :return: The email of this CreateUser.
        :rtype: str
        """
        return self._email

    @email.setter
    def email(self, email: str):
        """Sets the email of this CreateUser.


        :param email: The email of this CreateUser.
        :type email: str
        """

        self._email = email

    @property
    def password(self) -> str:
        """Gets the password of this CreateUser.


        :return: The password of this CreateUser.
        :rtype: str
        """
        return self._password

    @password.setter
    def password(self, password: str):
        """Sets the password of this CreateUser.


        :param password: The password of this CreateUser.
        :type password: str
        """

        self._password = password