# swagger_server/models/user.py
# coding: utf-8

from __future__ import absolute_import
from datetime import date, datetime  # noqa: F401

from typing import List, Dict  # noqa: F401

from swagger_server.models.base_model_ import Model
from swagger_server import util


class User(Model):
    """NOTE: This class is auto generated by the swagger code generator program.

    Do not edit the class manually.
    """

    def __init__(self, userid: str=None):  # noqa: E501
        """User - a model defined in Swagger

        :param userid: The userid of this User.  # noqa: E501
        :type userid: str
        """
        self.swagger_types = {
            'userid': str
        }

        self.attribute_map = {
            'userid': 'userid'
        }

        self._userid = userid

    @classmethod
    def from_dict(cls, dikt) -> 'User':
        """Returns the dict as a model

        :param dikt: A dict.
        :type: dict
        :return: The user of this User.  # noqa: E501
        :rtype: User
        """
        return util.deserialize_model(dikt, cls)

    @property
    def userid(self) -> str:
        """Gets the userid of this User.


        :return: The userid of this User.
        :rtype: str
        """
        return self._userid

    @userid.setter
    def userid(self, userid: str):
        """Sets the userid of this User.


        :param userid: The userid of this User.
        :type userid: str
        """

        self._userid = userid